+++ StudentVerse/database/migrations/2023_07_23_092948_create_mail__boxes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('Title');
             $table->string('Description');
             $table->string('Vote');
-            $table->integer('Tagged_Id');
+            $table->unsignedBigInteger('Tagged_Id');
             $table->foreign("Tagged_Id")->references("id")->on("tagged__boxes");
