+++ StudentVerse/database/migrations/2023_07_23_093051_create_complaint__boxes_table.php
@@ -16,11 +16,11 @@ return new class extends Migration
             $table->string('Title');
             $table->string('Description');
             $table->string('Vote');
-            $table->integer('User_Id');
+            $table->unsignedBigInteger('User_Id');
     
             $table->foreign("User_Id")->references("id")->on("userboxes");
 
-            $table->integer('Tagged_Id');
+            $table->unsignedBigInteger('Tagged_Id');
             $table->foreign("Tagged_Id")->references("id")->on("tagged__boxes");
