+++ StudentVerse/database/migrations/2023_07_23_093236_create_petition__boxes_table.php
@@ -16,13 +16,13 @@ return new class extends Migration
             $table->string('Subject');
             $table->string('Description');
             $table->string('Vote');
-            $table->integer('User_Id');
+            $table->unsignedBigInteger('User_Id');
 
             $table->foreign("User_Id")->references("id")->on("userboxes");
 
 
 
-            $table->integer('Tagged_Id');
+            $table->unsignedBigInteger('Tagged_Id');
             $table->foreign("Tagged_Id")->references("id")->on("tagged__boxes");
