+++ StudentVerse/database/migrations/2023_07_23_093327_create_question__boxes_table.php
@@ -15,12 +15,12 @@ return new class extends Migration
             $table->id();
             $table->string('Subject');
             $table->string('Description');
-            $table->integer('User_Id');
+            $table->unsignedBigInteger('User_Id');
            
             $table->foreign("User_Id")->references("id")->on("userboxes");
 
             
-            $table->integer('Tagged_Id');
+            $table->unsignedBigInteger('Tagged_Id');
 
             $table->foreign("Tagged_Id")->references("id")->on("tagged__boxes");
